Crate tor_persist

tor-persist

Persistent data storage for use with Tor.

This crate is part of Arti, a project to implement Tor in Rust.

For now, users should construct storage objects directly with (for example) FsStateMgr::from_path_and_mistrust(), but use them primarily via the interfaces of the StateMgr trait.

License: MIT OR Apache-2.0

Structs

Error

An error that occurred while manipulating persistent state.

FsStateMgr

Implementation of StateMgr that stores state as JSON files on disk.

Enums

ErrorSource

An underlying error manipulating persistent state.

Futureproof

A wrapper type for types whose representation may change in future versions of Arti.

JsonValue

Represents any valid JSON value.

LockStatus

A possible outcome from calling StateMgr::try_lock()

Traits

StateMgr

An object that can manage persistent state.

StorageHandle

A handle to a storage system that stores objects of a single type to a single location.

Type Definitions

DynStorageHandle

Type wrapper for a reference-counted dyn StorageHandle.